prediction-consumer/pyproject.toml
[tool.poetry]
name = "prediction-consumer"
version = "0.1.0"
description = ""
authors = ["Corentin Duchêne <corentin.duchene@epita.fr>"]

[tool.poetry.dependencies]
python = "3.8.16"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
kafka-python = "^2.0.2"
"discord.py" = "^2.1.0"
aiokafka = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
